Several minor fixes regarding reaction output and dynamic geometry
The reaction output colors were converted from floating to integer.

The function cellblender_utils.project_files_path() now references the
project directory ABOVE the enclosed mcell directory. The two dynamic
geometry addons were still using it as if it referenced the mcell
subdirectory. Changed to "mcell_files_path()".

The capsule_test addon is working with these changes, but there is still
a problem with the box_test (which uses older code).

These two tests are being reviewed to estimate the effort to make
dynamic geometry scriptable at the user level through the data model.
This would require a generic frame change handler that can read either
MDL geometry or a custom binary format. Such a frame change handler is
available in both of those test scripts, and it could be included into
CellBlender itself to provide the dynamic geometry visualization from
the MDL or a similarly generated binary format.

+++ cellblender_reaction_output.py
@@ -257,9 +257,9 @@ class MCELL_OT_plot_rxn_output_with_selected(bpy.types.Operator):
                             mol_color = mats.get(mol_mat_name).diffuse_color
                             #print("Molecule color = ", mol_mat.diffuse_color)
 
-                            mol_color_red = 255 * mol_color.r
-                            mol_color_green = 255 * mol_color.g
-                            mol_color_blue = 255 * mol_color.b
+                            mol_color_red = int(255 * mol_color.r)
+                            mol_color_green = int(255 * mol_color.g)
+                            mol_color_blue = int(255 * mol_color.b)
                             color_string = " color=#%2.2x%2.2x%2.2x " % (
                                 mol_color_red, mol_color_green, mol_color_blue)
